backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
geoalchemy2==0.14.6
alembic==1.13.1
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
redis==5.0.1
boto3==1.34.34
pytest==7.4.4
pytest-asyncio==0.23.4
httpx==0.26.0
